Created the first view for the registration
There are going to be three of these, this is just the first one. It is also not finished. I have just laid out the skeleton. It turned out harder than I expected, because I was getting used to Android Studio and had to watch a few tutorials. Hopefully the next views are going to take less time, since I already learned how to use the IDE a bit better.