Create DB_Design_v2
With Ethan we just met at the library and discussed the DB design. He had an idea to use 3 tables instead of 5, but after some brainstorming, we decided that 5 will allow us to do more things and will be easier to use later. There was a file in Documents/Other called DB_Design.txt. We created a new one called DB_Design_v2 in there and wrote down the details of the new improved db. It is based on the first one, but more detailed and has some minor changes.