Worked more on the poster
I wrote more info, played around with text size and positioning, and added an Our Team section at the bottom. The last one I did, because I noticed that all of last year's posters have something similar. Currently there is space for 1 or 2 images. However, I think that there is too much text, so in the next few days I will try to find time to go over it and remove any unnecessary info. Hopefully this will open some more space, so that we can put in 2 or 3 images. I also want to add the logos of the different languages, softwares, etc. that we used, because last year's posters have that info as well. We can add the logos of Java, Android Studio, TiDB, SQLite, chatGPT.